Implemented cub device reduce. Removed duplicate memory allocation on the host.

+++ FF/types.h
@@ -2,8 +2,8 @@
 #define _TYPES_H
 #include <thrust/complex.h>
 
-typedef float REAL;
-//typedef double REAL;
+//typedef float REAL;
+typedef double REAL;
 using GPUComplexD = thrust::complex<double>;
 using GPUComplexF = thrust::complex<float>;
 using GPUComplex = thrust::complex<REAL>;
